+++ src/main/java/com/alexfoglia/exam/IMongoRepository.java
@@ -1,7 +1,9 @@
 package com.alexfoglia.exam;
 
+import java.math.BigInteger;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface IMongoRepository extends MongoRepository<Book,Integer> {
+public interface IMongoRepository extends MongoRepository<Book,BigInteger> {
 
 }
